+++ changelog.md
@@ -1,3 +1,3 @@
 ### Changed
 
-* Made the first two mission rerolls be free, and updated translations accordingly.
+* Made the mission registry public and added custom mission events (Completed, RewardClaimed, Rerolled, Reset) for addon developers to subscribe to.